backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0
pydantic>=2.10,<3.0
pydantic-settings>=2.6,<3.0
email-validator>=2.2,<3.0
python-jose>=3.3,<4.0
bcrypt>=4.2,<5.0
aiosqlite>=0.20,<1.0
